refactor(session): rename persistence/ports.py to persistence/contracts.py (#5366)
