+++ core/agent_harness/session/persistence/contracts.py
