+++ core/agent_harness/session/persistence/jsonl_repo.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from typing import Any
 
 import core.agent_harness.session.persistence.paths as storage_paths
-from core.agent_harness.session.persistence.ports import CHAT_KINDS
+from core.agent_harness.session.persistence.contracts import CHAT_KINDS
 from core.agent_harness.session.persistence.wal_recovery import dangling_tool_intents
 from core.state.transcript_window import SESSION_SUMMARY_PREFIX
