+++ core/agent_harness/session/persistence/paths.py
@@ -12,7 +12,7 @@ import contextlib
 import json
 from pathlib import Path
 
-from core.agent_harness.session.persistence.ports import CHAT_KINDS
+from core.agent_harness.session.persistence.contracts import CHAT_KINDS
 
 _NAME_MAX_CHARS = 50
